# services/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/recommendations": {
      "get": {
        "operationId": "RecommendationsController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "recommendations"
        ]
      }
    },
    "/recommendations/feedback": {
      "post": {
        "operationId": "RecommendationsController_createFeedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "recommendations"
        ]
      }
    },
...
  },
  "info": {
    "title": "HealthIA API",
    "description": "Main API for users, authentication and AI orchestration.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}